Finished blog post about onthisday.com dataset
Knees weak, nights are heavy... I'm not finishing that. Anyways, it was another day with a messed up sleep schedule. I was on YouTube and stumbled across this video about Stephen Hawking. Me being curious, I clicked on it. Of course, I watched the whole video. However, something that caught me off guard was at the end where they mention how he died the same day Einstein died. There's something about it that ticked me off.

It went from watching a YouTube video to spending 3 hours scraping data to gather celebrity births and deaths. However, THAT TURNED INTO ME SCRAPING A MAJORITY OF CONTENT FROM onthisday.com

By now, I have a dataset and I have no idea what to do with it. So I become a reasonable and understanding man and upload it to Kaggle. But of COURSE, I HAD to upload it on GitHub

But wait... how do I explain my methods? OH WAIT, I HAVE A BLOG THAT NOONE READS THAT I CAN USE

Another hour and a half flies by and now we're here. A new blog post, a new dataset that I can use and I can finally answer the questions that I wanted to answer back around 2 in the morning. It's 7am and I've reached serenity...

... that is until I have to actually MINE and UNDERSTAND the damn text