Improved formatting and included some needed text
I read over the blog post and saw some small flaws. Oh well... it happens.

One of the more peculiar choices was to format my code as if it was typed by someone that is a in love with the <enter> key. I'll be honest and say that I hated how I had to scroll vertically to see important code.

The sad part is that there's a really long stretch of HTML code that needs to be scrolled vertically. About fixing that... hahahahha